Put command function in  ThemeKit for export
