+++ lib/cli.js
@@ -1,7 +1,7 @@
 #!/usr/bin/env node
 
 const install = require('./install');
-const command = require('./command');
+const {command} = require('./themekit');
 const parsedArgv = require('minimist')(process.argv.slice(2));
 
 if (parsedArgv._[0] === 'install') {
